+++ QRCoderDemo/Form1.cs
@@ -33,7 +33,7 @@ namespace QRCoderDemo
             string level = comboBox1.SelectedItem.ToString();
             QRCodeGenerator.ECCLevel eccLevel = (QRCodeGenerator.ECCLevel)(level == "L" ? 0 : level == "M" ? 1 : level == "Q" ? 2 : 3);
             QRCodeGenerator qrGenerator = new QRCodeGenerator();
-            QRCodeGenerator.QRCode qrCode = qrGenerator.CreateQrCode(textBoxQRCode.Text, eccLevel);
+            QRCoder.QRCode qrCode = qrGenerator.CreateQrCode(textBoxQRCode.Text, eccLevel);
 
             pictureBoxQRCode.BackgroundImage = qrCode.GetGraphic(20, Color.Black, Color.White, getIconBitmap());
         }
